create API endpoint that could be used
